+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -16,6 +16,7 @@ package com.vmware.xenon.common;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.UnsupportedEncodingException;
 import java.net.InetAddress;
 import java.net.NetworkInterface;
 import java.net.ServerSocket;
@@ -29,6 +30,7 @@ import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.security.GeneralSecurityException;
 import java.security.KeyStore;
+import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
@@ -66,7 +68,6 @@ import java.util.logging.ConsoleHandler;
 import java.util.logging.FileHandler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
 
@@ -123,7 +124,6 @@ import com.vmware.xenon.services.common.TransactionFactoryService;
 import com.vmware.xenon.services.common.UpdateIndexRequest;
 import com.vmware.xenon.services.common.UserGroupService;
 import com.vmware.xenon.services.common.UserService;
-import com.vmware.xenon.services.common.authn.AuthenticationConstants;
 import com.vmware.xenon.services.common.authn.BasicAuthenticationService;
 
 /**
@@ -258,6 +258,11 @@ public class ServiceHost implements ServiceRequestSender {
          * the JAR file of the host
          */
         public Path resourceSandbox;
+
+        /**
+         * Command line argument. Specifies the secret to use to sign and verify tokens.
+         */
+        public String tokenSecret;
     }
 
     private static final LogFormatter LOG_FORMATTER = new LogFormatter();
@@ -390,6 +395,7 @@ public class ServiceHost implements ServiceRequestSender {
         public boolean isPeerSynchronizationEnabled;
         public int peerSynchronizationTimeLimitSeconds;
         public boolean isAuthorizationEnabled;
+        public byte[] tokenSecret;
         public transient boolean isStarted;
         public transient boolean isStopping;
         public SystemHostInfo systemInfo;
@@ -604,12 +610,6 @@ public class ServiceHost implements ServiceRequestSender {
 
         updateSystemInfo(false);
 
-        // Create token signer and verifier
-        this.tokenSigner = new Signer(
-                AuthenticationConstants.JWT_SECRET.getBytes(Utils.CHARSET));
-        this.tokenVerifier = new Verifier(
-                AuthenticationConstants.JWT_SECRET.getBytes(Utils.CHARSET));
-
         // Set default limits for memory utilization on core services and the host
         if (getServiceMemoryLimitMB(ROOT_PATH, MemoryLimitType.EXACT) == null) {
             setServiceMemoryLimit(ROOT_PATH, DEFAULT_PCT_MEMORY_LIMIT);
@@ -632,8 +632,20 @@ public class ServiceHost implements ServiceRequestSender {
         return this;
     }
 
-    private void initializeStateFromArguments(File s, Arguments args) throws URISyntaxException {
+    private void prepareTokenSecret() {
+        if (this.state.tokenSecret != null) {
+            return;
+        }
 
+        log(Level.WARNING, "No token secret specified, using random string");
+
+        SecureRandom random = new SecureRandom();
+        this.state.tokenSecret = new byte[32];
+        random.nextBytes(this.state.tokenSecret);
+    }
+
+    private void initializeStateFromArguments(File s, Arguments args) throws
+            URISyntaxException, UnsupportedEncodingException {
         if (args.resourceSandbox != null) {
             File resDir = args.resourceSandbox.toFile();
             if (resDir.exists()) {
@@ -663,6 +675,10 @@ public class ServiceHost implements ServiceRequestSender {
         this.state.peerSynchronizationTimeLimitSeconds = args.perFactoryPeerSynchronizationLimitSeconds;
         this.state.isPeerSynchronizationEnabled = args.isPeerSynchronizationEnabled;
         this.state.isAuthorizationEnabled = args.isAuthorizationEnabled;
+
+        if (args.tokenSecret != null) {
+            this.state.tokenSecret = args.tokenSecret.getBytes(Utils.CHARSET);
+        }
 
         File hostStateFile = new File(s, SERVICE_HOST_STATE_FILE);
         String errorFmt = hostStateFile.getPath()
@@ -834,6 +850,13 @@ public class ServiceHost implements ServiceRequestSender {
             throw new IllegalStateException("Already started");
         }
         this.state.isAuthorizationEnabled = isAuthorizationEnabled;
+    }
+
+    public void setTokenSecret(byte[] tokenSecret) {
+        if (isStarted()) {
+            throw new IllegalStateException("Already started");
+        }
+        this.state.tokenSecret = tokenSecret;
     }
 
     public boolean isPeerSynchronizationEnabled() {
@@ -1081,7 +1104,6 @@ public class ServiceHost implements ServiceRequestSender {
     }
 
     private ServiceHost startImpl() throws Throwable {
-
         synchronized (this.state) {
             if (isStarted()) {
                 return this;
@@ -1093,6 +1115,10 @@ public class ServiceHost implements ServiceRequestSender {
         if (this.isAuthorizationEnabled() && this.authorizationService == null) {
             this.authorizationService = new AuthorizationContextService();
         }
+
+        prepareTokenSecret();
+        this.tokenSigner = new Signer(this.state.tokenSecret);
+        this.tokenVerifier = new Verifier(this.state.tokenSecret);
 
         this.executor = Executors.newWorkStealingPool(Utils.DEFAULT_THREAD_COUNT);
         this.scheduledExecutor = Executors.newScheduledThreadPool(Utils.DEFAULT_THREAD_COUNT,
@@ -3176,7 +3202,7 @@ public class ServiceHost implements ServiceRequestSender {
             if (cookies == null) {
                 return null;
             }
-            token = cookies.get(AuthenticationConstants.XENON_JWT_COOKIE);
+            token = cookies.get(Operation.REQUEST_AUTH_TOKEN_COOKIE);
         }
 
         if (token == null) {
@@ -5614,7 +5640,7 @@ public class ServiceHost implements ServiceRequestSender {
      */
     private AuthorizationContext createAuthorizationContext(String userLink) {
         Claims.Builder cb = new Claims.Builder();
-        cb.setIssuer(AuthenticationConstants.JWT_ISSUER);
+        cb.setIssuer(Claims.DEFAULT_ISSUER);
         cb.setSubject(userLink);
 
         // Set an effective expiration to never