+++ xenon-websocket-test/src/test/java/com/vmware/xenon/common/test/websockets/JsWebSocket.java
@@ -46,7 +46,6 @@ import io.netty.handler.codec.http.websocketx.WebSocketVersion;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
-
 import org.mozilla.javascript.Context;
 import org.mozilla.javascript.Function;
 import org.mozilla.javascript.NativeObject;
@@ -55,10 +54,10 @@ import org.mozilla.javascript.annotations.JSFunction;
 import org.mozilla.javascript.annotations.JSGetter;
 import org.mozilla.javascript.annotations.JSSetter;
 
+import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.OperationContext;
 import com.vmware.xenon.common.ServiceHost;
 import com.vmware.xenon.common.http.netty.CookieJar;
-import com.vmware.xenon.services.common.authn.AuthenticationConstants;
 
 /**
  * Lightweight partial implementation of JS Websocket API
@@ -213,7 +212,7 @@ public class JsWebSocket extends ScriptableObject {
         if (OperationContext.getAuthorizationContext() != null
                 && OperationContext.getAuthorizationContext().getToken() != null) {
             headers.add(HttpHeaderNames.COOKIE, CookieJar.encodeCookies(Collections.singletonMap(
-                    AuthenticationConstants.XENON_JWT_COOKIE, OperationContext.getAuthorizationContext().getToken())));
+                    Operation.REQUEST_AUTH_TOKEN_COOKIE, OperationContext.getAuthorizationContext().getToken())));
         }
         final WebSocketClientHandler handler = new WebSocketClientHandler(
                 WebSocketClientHandshakerFactory.newHandshaker(